requirements.txt
numpy
PyYAML
pillow
pip
setuptools